all AWS query handler tests are working!
